Add 3 files from home that replace 3 from work
This is an attempt to bring different versions created pre-GitHub into synchronization by uploading all versions to GitHub.  This commit will upload 3 files, Node.java, Tree.java, and TreeTraversals.java, which were stored on my home laptop.  It is my hope that GitHub will allow these files to be compared to the ones of the same name that are already in this repository